テナントをマネージャー配下に修正
