--- app/javascript/packs/tenants.js
@@ -1,21 +1,0 @@
-// This file is automatically compiled by Webpack, along with any other files
-// present in this directory. You're encouraged to place your actual application logic in
-// a relevant structure within app/javascript and only use these pack files to reference
-// that code so it'll be compiled.
-
-require("@rails/ujs").start()
-require("turbolinks").start()
-require("@rails/activestorage").start()
-require("channels")
-require('admin-lte');
-require("jquery");
-
-
-import 'bootstrap';
-import '../stylesheets/tenants';
-import "@fortawesome/fontawesome-free/js/all";
-
-document.addEventListener("turbolinks:load", () => {
-  $('[data-toggle="tooltip"]').tooltip()
-});
-
