+++ app/javascript/stylesheets/managers/tenants.scss
