+++ app/views/layouts/managers/_header.html.erb
@@ -9,49 +9,17 @@
     <li class="nav-item">
       <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
     </li>
-    <li class="nav-item d-none d-sm-inline-block">
-      <%= link_to "投稿した記事一覧", admins_dash_boards_path, class:"nav-link", method: :get %>
-    </li>
   </ul>
-  <!-- Right navbar links -->
-  <ul class="navbar-nav ml-auto">
-    <li class="dropdown">
-      <% if request.fullpath.include?("/admins/articles") || request.fullpath.include?("/admins/dash_boards") %>
-        <li class="nav-item">
-          <%= link_to new_admins_article_path, class: "nav-link", method: :get do %>
-            <p>記事投稿</p>
-          <% end %>
-        </li> 
-      <% elsif request.fullpath == "/admins/profiles" %>
-        <li class="nav-item">
-          <%= link_to new_admins_profile_path, class: "nav-link", method: :get do %>
-            <p>プロフィール投稿</p>
-          <% end %>
-        </li>
-      <% elsif request.fullpath == "/admins/posts" %>
-        <li class="nav-item">
-          <%= link_to new_admins_post_path, class: "nav-link", method: :get do %>
-            <p>動画投稿</p>
-          <% end %>
-        </li>
-      <% elsif request.fullpath == "/admins/tweets" %>
-        <li class="nav-item">
-          <%= link_to new_admins_tweet_path, class: "nav-link", remote: true, method: :get do %>
-            <p>つぶやき投稿</p>
-          <% end %>
-        </li>
-      <% end %>
-    </li>    
-  </ul>
-  <div class="dropdown navbar-nav" style="padding-bottom:6px">
+  
+  <div class="dropdown navbar-nav ml-auto" style="padding-bottom:6px">
     <div class="nav-link" id="dropdownMenuButton button_group" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-    <%= image_tag "admin.png", size: "30x30" %><b class="caret"></b>
-  </div>
+      <%= image_tag "admin.png", size: "30x30" %><b class="caret"></b>
+    </div>
     <ul class="dropdown-menu dropdown-menu-right dropdown-menu-header">
-      <li style="padding-left:3px"><%= current_admin.name %></li>
+      <li style="padding-left:3px"><%= "マネージャーの名前" %></li>
       <hr style="height:0.5px">
-      <li style="padding-left:3px"><%= link_to "詳細画面", admins_show_admins_profiles_path, method: :get %></li>
-      <li style="padding-left:3px"><%= link_to "ログアウト", destroy_user_session_path, method: :delete %></li>
+      <li style="padding-left:3px"><%= link_to "詳細画面", "#" , method: :get %></li>
+      <li style="padding-left:3px"><%= link_to "ログアウト", destroy_manager_session_path, method: :delete %></li>
     </ul>
   </div>
 </nav>
