--- app/views/layouts/tenants.html.erb
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>App</title>
-    <meta name="viewport" content="width=device-width,initial-scale=1">
-    <%= csrf_meta_tags %>
-    <%= csp_meta_tag %>
-
-    <%= stylesheet_pack_tag 'tenants', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_pack_tag 'tenants', 'data-turbolinks-track': 'reload' %>
-    <script type="text/javascript" async src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-MML-AM_CHTML"></script>
-    <script type="text/x-mathjax-config">
-      MathJax.Hub.Config({
-        tex2jax: {
-          inlineMath: [["$","$"], ["\\(","\\)"]],
-          displayMath: [ ['$$','$$'], ["\\[","\\]"] ]
-        }
-      });
-    </script>
-  </head>
-
-  <body class="hold-transition sidebar-mini" id="<%= page_body_id(params) %>">
-    <% if params[:controller] != "managers/sessions" %>
-      <div class="wrapper">
-        <%= render "layouts/tenants/header" %>
-        <%= render "layouts/tenants/side" %>
-        <div class="content-wrapper">
-          <%= render "layouts/flash" %>
-          <!-- Content Header (Page header) -->
-          <div class="content-header">
-            <div class="container-fluid">
-              <div class="row mb-2">
-                <div class="col-sm-6">
-                  <h1 class="m-0"><%= yield(:title) %></h1>
-                </div><!-- /.col -->
-              </div><!-- /.row -->
-            </div><!-- /.container-fluid -->
-          </div>
-          <%= render template: "layouts/application" %>
-          <%= debug(params) if Rails.env.development? %>
-        </div>
-        <%= render "layouts/tenants/control_side" %>
-        <%= render "layouts/tenants/footer" %>
-      </div>
-    <% else %>
-      <div class="content">
-        <%= render "layouts/flash" %>
-        <!-- Content Header (Page header) -->
-        <div class="content-header">
-          <div class="container-fluid">
-            <div class="row mb-2">
-              <div class="col-sm-6">
-                <h1 class="m-0"><%= yield(:title) %></h1>
-              </div><!-- /.col -->
-            </div><!-- /.row -->
-          </div><!-- /.container-fluid -->
-        </div>
-        <%= render template: "layouts/application" %>
-      </div>
-    <% end %>
-  </body>
-</html>
