--- app/views/layouts/tenants/_footer.html.erb
@@ -1,8 +1,0 @@
-<footer class="main-footer">
-  <!-- To the right -->
-  <div class="float-right d-none d-sm-inline">
-    Anything you want
-  </div>
-  <!-- Default to the left -->
-  <strong>Copyright &copy; 2014-2021 <a href="https://adminlte.io">AdminLTE.io</a>.</strong> All rights reserved.
-</footer>
