--- app/views/layouts/tenants/_header.html.erb
@@ -1,25 +1,0 @@
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
-
-<!-- Navbar -->
-<nav class="main-header navbar navbar-expand navbar-white navbar-light">
-  <!-- Left navbar links -->
-  <ul class="navbar-nav">
-    <li class="nav-item">
-      <a class="nav-link" data-widget="pushmenu" href="#" role="button"><i class="fas fa-bars"></i></a>
-    </li>
-  </ul>
-  
-  <div class="dropdown navbar-nav ml-auto" style="padding-bottom:6px">
-    <div class="nav-link" id="dropdownMenuButton button_group" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-      <%= image_tag "admin.png", size: "30x30" %><b class="caret"></b>
-    </div>
-    <ul class="dropdown-menu dropdown-menu-right dropdown-menu-header">
-      <li style="padding-left:3px"><%= "マネージャーの名前" %></li>
-      <hr style="height:0.5px">
-      <li style="padding-left:3px"><%= link_to "詳細画面", "#" , method: :get %></li>
-      <li style="padding-left:3px"><%= link_to "ログアウト", destroy_manager_session_path, method: :delete %></li>
-    </ul>
-  </div>
-</nav>
