--- app/views/layouts/tenants/_side.html.erb
@@ -1,41 +1,0 @@
-<head><script src="https://kit.fontawesome.com/c3f6f3c4db.js" crossorigin="anonymous"></script></head>
-<aside class="main-sidebar sidebar-dark-primary elevation-4">
-  <!-- Brand Logo -->
-  <%= link_to "#", class:"brand-link", method: :get do %>
-    <%= image_tag "AdminLTELogo.png", class:"brand-image img-circle elevation-3", style: "opacity: .8" %>
-    <span class="brand-text font-weight-light">TecPutt</span>
-  <% end %>
-  <!-- Sidebar -->
-  <div class="sidebar">
-    <!-- Sidebar user panel (optional) -->
-    <div class="user-panel mt-3 pb-3 mb-3 d-flex">
-      <div class="image">
-      </div>
-      <div class="info">
-      </div>
-    </div>
-    <!-- Sidebar Menu -->
-    <nav class="mt-2">
-      <ul class="nav nav-pills nav-sidebar flex-column" data-widget="treeview" role="menu" data-accordion="false">
-        <li class="nav-item">
-          <%= link_to tenants_path, class:"nav-link", method: :get do %>
-            <i class="nav-icon fas fa-th"></i> <!-- <i class="nav-icon fas fa-newspaper"></i> user使用のアイコン -->
-            <p>
-              テナント一覧
-            </p>
-          <% end %>
-        </li>
-        <li class="nav-item">
-          <%= link_to "#", class:"nav-link", method: :get do %>
-            <i class="nav-icon far fa-plus-square"></i>
-            <p>
-              テナント登録
-            </p>
-          <% end %>
-        </li>
-      </ul>
-    </nav>
-    <!-- /.sidebar-menu -->
-  </div>
-  <!-- /.sidebar -->
-</aside>
