+++ app/views/managers/tenants/index.html.erb
@@ -45,3 +45,7 @@
             </div>
           </div>
         </div>
+      </div>
+    </div>
+  </div>
+ </div>
